scripts/requirements.txt
autopep8==1.5.4
click==7.1.2
pycodestyle==2.6.0
PyYAML==5.3.1
toml==0.10.1
